// 9gic-mon/src/lib/api/response.ts
export interface SearchedListItem {
  backgroundImageUrl: string;
  coUserPostCopname: string;
  coUserPostImageUrl: string;
  context: string;
  deadLine: string;
  etc: string;
  grade: number;
  mainWork: string;
  mainWorkStack: string;
  region: string;
  simpleIntro: string;
  subject: string;
  title: string;
  welfare: string;
}